apps/api/pyproject.toml
[project]
name = "music-autodl-api"
version = "0.1.0"
description = "FastAPI + Celery backend for music autodl"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.111.0",
  "uvicorn[standard]==0.30.0",
  "pydantic==2.8.2",
  "pydantic-settings==2.4.0",
  "SQLAlchemy==2.0.32",
  "psycopg2-binary==2.9.9",
  "alembic==1.13.2",  # для будущих миграций
  "httpx==0.27.0",
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "celery==5.4.0",
  "redis==5.0.7",
]

[tool.uvicorn]
reload = false
